Silenziato un warning

+++ libTPool/benchmarks/fillingBenchmarks.cpp
@@ -28,7 +28,7 @@ static void fillHistogramsAsync(benchmark::State &state) {
 
     std::for_each(v.begin(), v.end(), [&](const auto &arr) {
       std::for_each(std::execution::par_unseq, arr.begin(), arr.end(),
-                    [&](const auto &item) { histograms[&item - arr.data()].Fill(item); });
+                    [&](const auto &item) { histograms[static_cast<long unsigned int>(&item - arr.data())].Fill(item); });
     });
   }
 }
